Added Docker, ebpf (bpftrace) backend, started tests.
