+++ src/secimport/backends/common/instrumentation_backend.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class InstrumentationBackend(Enum):
+    DTRACE = "dtrace"
+    EBPF = "ebpf"  # bpftrace
