+++ src/secimport/templates/bpftrace/actions/log_file_system.bt
@@ -1,0 +1,1 @@
+printf("\t\t(TOUCHING FILESYSTEM): %s(%d) in python module %s\r\n",  @sysname[args->id], arg1,  @globals["current_module"]);
