+++ src/secimport/templates/bpftrace/actions/log_network.bt
@@ -1,0 +1,1 @@
+printf("\t\t(NETWORKING): %s in python module %s\r\n", @sysname[args->id], arg1, @globals["current_module"]);
