+++ src/secimport/templates/bpftrace/actions/log_python_module_entry.bt
@@ -1,0 +1,1 @@
+printf("%s, %s, depth=%d\n", str(arg0), str(arg1), @["depth"]) ;
