+++ src/secimport/templates/bpftrace/filters/file_system.bt
@@ -1,0 +1,1 @@
+if (@sysname[args->id] == "open" || @sysname[args->id] == "write")
