+++ src/secimport/templates/bpftrace/filters/networking.bt
@@ -1,0 +1,1 @@
+    if (@sysname[args->id] == "socket")
